Delete property "id" in User model
This propeprty is already defined by a superclass of user (with a different implementation).
Furthermore, the guarantee that the id of an user is unique among all existing
users is sufficient for the use case of the id and therefore no different (meaning
different from both the perevious implementation and the implementation in the superclass)
is needed.