Fix problem with repeated loading of Pyod algorithms
For loading the Pyod algorithms, the pyodtodb command is reused
(in parts, not very sofisticated, but it works for the testing purposes).
In the respective file, there is a attrubute which stores the paths to all Pyod algorithms.
When the Pyod algorithms are renamed, so that the requirements to import them into the database are met,
this attribute is also changed, so that the paths are up to date.
But this creates a problem, as the second time the Pyod algorithms are imported into the database
the paths are no longer up to date, because the files of the algorithms have been replaced by fresh ones.
This is solved by adding a attribute, to store the original data, and replacing the changed attribute
with the original data before the Pyod algorithms are imported a second time.
(Not very nice, but for testing purposes it works fine.)